Method deleteUser created

+++ src/main/java/com/youtube/demo/dao/UserRepository.java
@@ -4,9 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.youtube.demo.model.User;
 
-public interface UserRepository extends JpaRepository<User,Long>{
+public interface UserRepository extends JpaRepository<User,Integer>{
 	
 	@SuppressWarnings("unchecked")
 	User save(User user);
+
+	
 
 }
